requirements.txt
torchvision>=0.15
Pillow>=9.0
imageio>=2.28
imageio-ffmpeg>=0.4
numpy>=1.24
tqdm>=4.65
